requirements.txt
asgiref==3.7.2
Django==4.2.3
Pillow==10.0.0
psycopg==3.1.9
psycopg2-binary==2.9.6
sqlparse==0.4.4
typing_extensions==4.7.1
whitenoise==6.5.0
